% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGround_class.R
\name{check_WGround}
\alias{check_WGround}
\title{The validity function for class WGround}
\usage{
check_WGround(object)
}
\arguments{
\item{object}{an object puportedly of class \code{WGround}}
}
\value{
TRUE if valid, a character vector error messages if not
}
\description{
The validity function for class WGround
}
\keyword{internal}
