% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGround_class.R
\name{Parseconditiondata}
\alias{Parseconditiondata}
\title{A function to get current conditions data from wunderground
gets called from CurrentWeatherbyPWS method
Return list with conditions info. Throws an error if invalid input.}
\usage{
Parseconditiondata(url, local = F)
}
\arguments{
\item{url}{fully formed url for API call}

\item{local}{set to F for online mode otherwise T}
}
\value{
conditiondata list with selected elements
}
\description{
A function to get current conditions data from wunderground
gets called from CurrentWeatherbyPWS method
Return list with conditions info. Throws an error if invalid input.
}
