% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGround_class.R
\name{ParseHistoricalWeatherdata}
\alias{ParseHistoricalWeatherdata}
\title{A function to get historical weather data from wunderground
gets called from getHistoricalWeather method
Return data frame with historical weather data. Throws an error if query is unsuccessful.}
\usage{
ParseHistoricalWeatherdata(URL, Dt, PWSid, verbose, local = F)
}
\arguments{
\item{URL}{fully formed url for API call}

\item{Dt}{date of historical data}

\item{PWSid}{pws_id for the station}

\item{verbose}{shows details on process}

\item{local}{set to F for online mode otherwise T}
}
\value{
wdata_df data frame with selected columns
}
\description{
A function to get historical weather data from wunderground
gets called from getHistoricalWeather method
Return data frame with historical weather data. Throws an error if query is unsuccessful.
}
