% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGround_class.R
\name{ParsePWSdata}
\alias{ParsePWSdata}
\title{A function to get geo data from wunderground
gets called from getPWSinfo method
Return data frame with PWS info. Throws an error if invalid input.}
\usage{
ParsePWSdata(URL, local = F)
}
\arguments{
\item{URL}{fully formed url for API call}

\item{local}{set to F for online mode otherwise T}
}
\value{
pws_df data frame with selected columns
}
\description{
A function to get geo data from wunderground
gets called from getPWSinfo method
Return data frame with PWS info. Throws an error if invalid input.
}
