% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwsdata.R
\docType{package}
\name{rpwsdata}
\alias{rpwsdata}
\alias{rpwsdata-package}
\title{A package using R to get and analyze data from Wunderground}
\description{
Contains S4 class WGround and Shiny interface to query and
visualize data
}
\references{
Hadley Wickham's book at \url{http://r-pkgs.had.co.nz/}
}
\author{
Siddhartha Jetti - 50\% & Atul Lal - 50\%
}
