% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyoffline.R
\name{runShinyoffline-function}
\alias{runShinyoffline-function}
\alias{runShinyoffline}
\title{runShiny function to start shiny app in offline mode.}
\usage{
runShinyoffline()
}
\description{
A function to start shiny app for offline use
}
