% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGround_class.R
\name{WGround}
\alias{WGround}
\title{WGround class constructor function}
\usage{
WGround(APIkey = "", city = "Fremont", state = "CA",
  runmode = "offline")
}
\arguments{
\item{APIkey}{API key from Wunderground}

\item{city}{city}

\item{state}{state or country for international city}

\item{runmode}{online or offline, default: online}
}
\description{
WGround class constructor function
}
\examples{
obj <- new("WGround", city="Fremont",state= "CA",runmode="offline")

}
